+++ src/core/automix/SharedAudioContext.ts
@@ -1,9 +1,16 @@
-import type { IExtendedAudioContext } from "./BaseAudioPlayer";
+import type { IExtendedAudioContext } from "@/types/audio/context";
 
+/** 共享音频上下文 */
 let sharedContext: IExtendedAudioContext | null = null;
+/** 主输入节点 */
 let masterInput: GainNode | null = null;
+/** 主限制器节点 */
 let masterLimiter: DynamicsCompressorNode | null = null;
 
+/**
+ * 获取共享音频上下文
+ * @returns 共享音频上下文
+ */
 export const getSharedAudioContext = (): IExtendedAudioContext => {
   if (!sharedContext) {
     const AudioContextClass =
@@ -18,6 +25,10 @@ export const getSharedAudioContext = (): IExtendedAudioContext => {
   return sharedContext;
 };
 
+/**
+ * 获取主输入节点
+ * @returns 主输入节点
+ */
 export const getSharedMasterInput = (): GainNode => {
   const ctx = getSharedAudioContext();
   if (!masterInput) {
@@ -36,6 +47,10 @@ export const getSharedMasterInput = (): GainNode => {
   return masterInput;
 };
 
+/**
+ * 获取主限制器节点
+ * @returns 主限制器节点
+ */
 export const getSharedMasterLimiter = (): DynamicsCompressorNode | null => {
   return masterLimiter;
 };
