+++ src/types/audio/context.ts
@@ -1,0 +1,4 @@
+/** 扩充 AudioContext 接口以支持 setSinkId (实验性 API) */
+export interface IExtendedAudioContext extends AudioContext {
+  setSinkId(deviceId: string): Promise<void>;
+}
